requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==3.2.2
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
twilio==8.10.0
python-dotenv==1.0.0
email-validator==2.1.0